Using the spigot API to send action bar messages
